broke out filterselect to separate type & restored old select
