--- src/Form/types/Select.styled.js
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-
-export const Input = styled.input`
-    &:disabled {
-        color: black!important;
-    }
-`;
-
-export const SelectFilter = styled.div`
-    display: flex;
-    width: 100%;
-`;
